Docs/releases: Finalize 4.22/4.22.01 release notes
Now that the 4.22 release tag has been added to git, update the release
notes with the final statistics and wording.

We also decided to add a fix submitted immediately after the 4.22
release was tagged into the release package and do a point release.

This also adds an expected date for the next release

Signed-off-by: Martin Roth <gaumless@gmail.com>
Change-Id: Iae9653a275fcc1d11efbb88e12676f332be0a5dc
Reviewed-on: https://review.coreboot.org/c/coreboot/+/79147
Reviewed-by: Felix Singer <service+coreboot-gerrit@felixsinger.de>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>