Revert "nipperkin: Fix WLAN to GEN2 speed" & "Disable PSPP for WLAN"
Updated Linux FW works with PCI gen3 speed and PSPP.

This reverts
commit 05c9a850fd21 ("mb/google/nipperkin: Fix WLAN to GEN2 speed")
https://review.coreboot.org/c/coreboot/+/63593

and
commit 76fddd963925 ("mb/google/nipperkin: Disable PSPP for WLAN")
https://review.coreboot.org/c/coreboot/+/63722

The changes are overlapped and are reverted together.

BUG=b:240426142 & b:228830362

The system is able to ran over 2500 cycles on Nipperkin with command
suspend_stress_test -c 10000 --wake_min 10 --suspend_min 10 \
    --nofw_errors_fatal

The whole variant_update_dxio_descriptors is empty and is pushed back
to weak function.

Change-Id: Id207076542edc8ea0cabc6e02e29856c2b6803c7
Signed-off-by: Zheng Bao <fishbaozi@gmail.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/79172
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Tim Van Patten <timvp@google.com>